requirements.txt
attrs==24.2.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.58.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
onnxruntime==1.20.1
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.0
PSNAWP==2.2.0
pycountry==24.6.1
PyMatting==1.1.14
pyparsing==3.2.3
pyrate-limiter==2.10.0
python-dateutil==2.9.0.post0
python-steam-api==2.1.1
pytz==2025.2
rawgpy==1.0.3
referencing==0.36.2
rembg==2.0.60
requests==2.32.3
requests-ratelimiter==0.7.0
rpds-py==0.25.0
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
soupsieve==2.7
sympy==1.14.0
tifffile==2025.5.10
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
